Create storage pool ahead of virt-install.
- noticed with Fedora26 or a more recent version of libvirt that if the storage pool doesn't exist before running virt-install, I started getting the following error:

Error: --disk baz.qcow2,format=qcow2,bus=virtio: 'vols'

In debug mode, the tail end of the traceback was:

  File "/usr/share/virt-manager/virtinst/connection.py", line 252, in _cache_new_pool_raw
    vollist = self._fetch_cache[self._FETCH_KEY_VOLS]
KeyError: 'vols'

Seems like precreating the storage pool fixes this problem.  Without it, running the script a second time would work because the storage pool was created on the first go.

- other refactoring