Resume interrupted downloads
Currently, if an image download is interrupted or fails, a partial file
is left in the image directory and a subsequent kvm-install-vm create
command will blissfully use the partial image, which then silently fails
to create the guest.

Instead, name the files with the '.part' file extension while the file
download is in progress, then pivot to the real image name after the
download is complete.  Use the wget '--continue' option to resume
partial downloads so we don't have to restart the download from the
beginning.