Fix kvm-install-vm remove -v flag handling
Fix the handling of flags in the remove subcommand. Add the missing
shift operator so the guest name is properly detected. Use the internal
destroy_vm when overritting a guest during the kvm-install-vm create
command, since we should not perform the shift in that code path.

This commit fixes the error when a flag is given to the remove command.

    $ kvm-install-vm remove -v test01
    Please specify a single host to remove.
    Run 'kvm-install-vm help remove' for usage.