Eject the CI CDROM by image path name (#36)
Currently the eject command hard-codes device name 'hda' to eject the CI
CDROM image. However the device name depends on the distribution of the
kvm host system.

Fortunately, the 'virsh change-media' command can alternately use the
path of the image instead of the device name, so use that instead of the
hard-coded device name 'hda' to eject the CDROM image.