Add command wrapper for verbose mode
Add a wrapper function to automatically print command arguments when
running in verbose mode. Currently, this is only done for the `virsh
pool-create-as` and `virt-install` commands, but more commands could
be added in the future.

Having the wrapper run the command removes the redundant command
arguments and will allow us to more easily support a variable number
of virt-install arguments

This commit introduces the following shell helper functions:

    join             Join strings into a single string, joined
                     by a delimiter.

    output_command   Print the command arguments for verbose mode,
                     formatted as one argument per line, with line
                     continuation characters as needed.

    run              Command execution wrapper to print the command
                     when running in verbose mode and to redirect
                     output to the current log file.