Hypervisor custom parameters
Support custom virt-install parameters for advanced usage. This allows
for advanced customization and to support Xen and other non-KVM
hypervisors.

Users may specify hypervisor specific settings in the $HOME/.kivrc
customization file, without resorting to patching kvm-install-vm.

With this change, none of the virt-install parameters are hardcoded in
the script, values may be specified to be empty to avoid setting a given
parameter, and additional options may be provided as needed.

This commit adds the following custom virt-install options:

    NETWORK_MODEL       --network model value (default: virtio)
    NETWORK_EXTRA       extra --network options
    DISK_BUS            --disk bus type (default: virtio)
    DISK_EXTRA          extra --disk parameters
    CI_ISO_DEVICE       ci cdrom device type (default: cdrom)
    CI_ISO_EXTRA        extra ci cdrom parameters (default: "")
    GRAPHICS_LISTEN     --graphics listen value (default: localhost)
    GRAPHICS_EXTRA      extra graphics parameters
    VIRT_INSTALL_EXTRA  extra virt-install options

The following virt-install options are dependent on the image being
installed, and so are determined the value of -t command line option.

    OS_TYPE             --os-type (currently always "linux")
    DISK_FORMAT         --disk format (currently always "qcow2")