Check os-variant before starting vm
Check the osinfo-db is not too old, and if so issue an error message to
explain how to upgrade the osinfo-db.

The osinfo-db provided by package managers tend to lag the upstream
libosinfo project. This makes it harder to support new os versions when
your kvm is running on something like CentOS or Ubuntu LTS.  Fortunately
the libosinfo project provides downloads and a tool to import the
osinfo-db without impacting the files managed by the package manager,
which makes using kvm-import-vm create a much more pleasant experience.